first re-read

+++ src/Bisection.h
@@ -7,27 +7,26 @@
 
 #include "Equation_Solver.h"
 
-/** \brief Daughter class for the bisection method for non liner equation solving
+/** \brief Daughter class for the bisection method for 1D non linear equation solving
   * This class implements the bisection solving method for non linear equation of type
   * f(x) = 0, with f and x 1D only
   */
 class Bisection : public Equation_Solver {
-// the function needs to be continue entre a et b et f(a)<0 et f(b)>0
-
 private:
-    double b;
+    // the function needs to be continuous between a and b, with f(a)<0 and f(b)>0
+    double b; // x0 from the mother class will be used as a, the left starting point, b is the second starting point
 
 public:
     /// Constructors
     Bisection();
     Bisection(double (*fx)(double x));
-    Bisection(double (*fx)(double x), double starting_point, double b, double iter, double tol);
-    Bisection(double (*fx)(double x), bool acc, double starting_point, double b, double iter, double tol);
+    Bisection(double (*fx)(double x), double starting_point, double b, unsigned int iter, double tol);
+    Bisection(double (*fx)(double x), bool acc, double starting_point, double b, unsigned int iter, double tol);
 
     /// Desctructor
     ~Bisection();
 
-    /** \brief Overrided solving method
+    /** \brief Overriden solving method
       * This method implements the bisection solving algorithm
       * @return the guessed root
       */
