+++ src/Equation_Solver.h
@@ -8,22 +8,22 @@
 #include "Mama_Solver.h"
 
 
-/** \brief 2nd line super class for any king of methods for non liner equation solving
+/** \brief Pure virtual class, inheriting from Mama_Solver, for any king of methods for simple 1D non liner equation solving
   * This class is for every possible solving methods for equation solving of type
   * f(x) = 0, with f and x 1D only
   */
 class Equation_Solver : public Mama_Solver {
 protected :
-    double x0;
-    double (*function)(double);
-    bool acceleration;
+    double x0; // initial starting point, gets modified at each iteration to become the best guess
+    double (*function)(double); // equation to solve f(x) = 0
+    bool acceleration; // whether to use Aitken acceleration or not
 
 public :
     /// Constructors
     Equation_Solver();
     Equation_Solver(double (*fx)(double x));
-    Equation_Solver(double (*fx)(double x), double starting_point, double iter, double tol);
-    Equation_Solver(double (*fx)(double x), bool acc, double starting_point, double iter, double tol);
+    Equation_Solver(double (*fx)(double x), double starting_point, unsigned int iter, double tol);
+    Equation_Solver(double (*fx)(double x), bool acc, double starting_point, unsigned int iter, double tol);
     /// Destructor
     virtual ~Equation_Solver();
 
@@ -34,7 +34,7 @@ public :
     virtual double Solve() = 0;
 
     /** \brief Pretty method to compute and print the results
-      * This method computes and displays the results in a clear and comprehensive way
+      * This inherited method computes and displays the results in a clear and comprehensive way
       */
     void Result() final;
